grammar: Fix BooleanLiteral rule
Due to how "one of" rules worked the BooleanLiteral rule was exactly
"true false" when it should have been "true" and "false", independently.

The "one of" grammar modifier only applies to TOKENS rules which are
rules that are specifically not found in the productions. Because of
this, the "one of" semantic of having every token on a few lines with
space delimiters did not work.