scanner,parser: Add support for comments (single, multi, and javadoc)
This adds support for 3 types of comments:

- Single line. Prefixed with "//" until the end of the line.
- Multi-line. Contained by "/*" and "*/" among multiple lines.
- Javadoc. Contained by "/**" and "*/" among multiple lines.

In order to do this, the parser needs to filter out any comment lexemes
as they are not accounted for in the grammar. In addition, the input
scanned was changed to Bytes (a Slice(UInt8)), which is much safer as it
is read-only with better bounds checks.