parser: Add support for parsing empty files.
Previously, empty files would fail during parsing due to an obscure
rule. This was because "CompilationRule" consists of an empty RHS, which
incorrectly read as a RHS of size 1, as a result of how the LALR1
prediction table was created.

This was fixed, along with a basic test to check this.