build: update makefile to allow assignment configurations, including zip
The tl;dr is the makefile now requires `ASSN=Ax` parameter during
compilation to compile in a default configuration, which is just the
default end stage for now. This is just passing forward the make
parameter as a compiler flag.

It also handles `orangejoos.zip` as it will write the parameter to a
file that is then imported on marmoset for the configuration. This lets
you do `make orangejoos.zip ASSN=A1` to generate an A1 zip.