Fix build and tester
After modifying the make system, the tester was not working. This fixes
that. The tester builds with no assignment target, but is expected to
define the end stage. In the future, the tester script could have a
switch of assignments and the test folders so the proper make target is
used.