nameres: resolve root packages and qualified names
If a source file does not declare a package, it is considered part of a
special root package. As normal, all protected classes are visible
within the same package.

Another change was to allow the lookup of qualified names. If a
reference is made to `foo.Bar`, the type does not need to be imported at
all.