change instanceof expression to be specific
This changes simplification to not convert instanceof expression to
`ExprOp` but instead a specialized `ExprInstanceOf`. This was because
the InstanceOf RHS is not an expression, but a type. Because of this, it
will behave differently. It will also differ significantly in code
generation due to how runtime typing will work.

Making this change also opened up the ability to remove `Typ` from being
an `Expr`. That extension was only required as a workaround for
InstanceOf.

(Sidenote: Added the ClassTypResolution visitor to test that
 instanceof type resolving was working :) )