fix CastExpr AST generation
This changes the `CastExpr`. Previously to represent the cast type it
contained a `Name`, `Expr`, or `PrimitiveTyp`. This was simplified to
only a `Typ`.

This was done in order to fix problems that prevented resolving the
`ReferenceTyp` names during name resolution. It also vastly simplifies
how casts are handled.

The visitor checking the cast validity was also removed, and the check
is now done during simplification. It will ensure the child expression
tree of the CastExpr is a Name, and if not an error will be thrown. The
name is also ensured to be valid during name resolution.