change ReferenceTyp to ClassTyp
This change properly reflects what it represents. In the grammar, and in
Java, a Reference type is one that is passed by reference and handled as
a reference. This is why a reference type includes arrays, not just
user-defined types.

ReferenceTyp will come up again during code generation, but only for
that sub-system.