add expression QualifiedName disambiguation (into FieldAccess)
Before, if a `FieldAccess` only involved local variables and no other
type of expressions, it was ambiguously parsed as a `QualifiedName`.

Inside expressions, the only time a `QualifiedName` should occur inside
of an `ExprRef` is if it is a variable access and a string of field
accesses. The visitor now converts these prior to variable resolution.