nameres: prevent static field references to non-static fields
Previously, a static field initializer could refer to a non-static
field. This behaviour would change depending on whether or not a static
or non-static method was before the field declarations.

Now, when entering the context of a field initializer the appropriate
namespace is used (static vs. non-static).

This now passes `Je_16_StaticThis_StaticFieldInitializer`.