Fix MethodEnvironmentVisitor must visit interfaces
Previously, there was an issue where seemingly any time we visit a
MethodDecl that is the child of an Interface, before a ClassDecl has
been visited, an error would be thrown.
This was because the MethodDecl visitor assumed that a class has
been visited already.

The code has been modifide to visit all TypeDecls instead. This was done
because we still want to visit the MethodDecls inside of Interfaces to
check their names and the names of their parameters.