typing: fix protected field cases
Checking if protected access is allowed is slightly more complicated
than expected. Consider the following cases, where all classes are in
separate packages:

    class A
      protected field foo;

    class B extends A

    class C extends B
      foo(D d) { return d.foo; }
      foo(E e) { return e.foo; }
      foo(F f) { return f.foo; }

    class D extends C

    class E extends B

    class F extends C
      protected field foo;

In this example, `foo(D d)` is allowed, even though C is not in the same
package as A. This is because C is a sub-class of D. On the other
hand, `foo(E e)` generates a compile time error because even though the
field is declared in a class that is in the sub-class to C because E is
not a sub-class of C. `foo(F f)` also fails because the field being
referenced is not in a sub-class of `C`.