parser: add parse tree flattening step
This will reduce the depth of the parse tree for expressions. Due to how
the grammar is designed for handling expression precedence, parsing
expressions caused an large tree depth. An iterative based solution was
added to traverse through the tree and flatten it.

Ideally, the entire simplification process will be translated to an
iterative solution similar to the flattening step in order to prevent
recursive depth issues (due to toom any function calls).